failed test fix; modify package, application settings; apply google java style;
