+++ payment/clients/build.gradle
@@ -8,8 +8,8 @@ plugins {
 }
 
 version = project.version
-group = 'com.vroong.template'
-archivesBaseName = 'template-api-java-client'
+group = 'com.vroong.payment'
+archivesBaseName = 'payment-api-java-client'
 def specPath = "${parent.projectDir}/src/main/resources/swagger/api.yml"
 
 sourceCompatibility = '1.8'
@@ -30,13 +30,13 @@ openApiGenerate {
     generatorName = "java"
     inputSpec = specPath
     outputDir = "${buildDir}/java-client"
-    apiPackage = "com.vroong.template.rest"
-    modelPackage = "com.vroong.template.rest"
+    apiPackage = "com.vroong.payment.rest"
+    modelPackage = "com.vroong.payment.rest"
     apiFilesConstrainedTo = [""]
     modelFilesConstrainedTo = [""]
     supportingFilesConstrainedTo = [""]
     configOptions = [
-            title: "template",
+            title: "payment",
             useTags: "true",
             dateLibrary: "java8",
             java8: "true",
