+++ payment/src/main/java/com/vroong/payment/adapter/in/rest/Fixtures.java
@@ -1,0 +1,19 @@
+package com.vroong.payment.adapter.in.rest;
+
+import com.vroong.payment.domain.PaymentState;
+import com.vroong.payment.rest.MoneyDto;
+import com.vroong.payment.rest.PaymentDto;
+import com.vroong.payment.rest.PaymentStateDto;
+import java.math.BigDecimal;
+
+public class Fixtures {
+
+  static PaymentDto aPaymentDto() {
+    return new PaymentDto()
+        .paymentId(1L)
+        .cardNumber("123456789101234")
+        .approvalNumber("123456789101111")
+        .amount(new MoneyDto().value(new BigDecimal(100000)))
+        .state(PaymentStateDto.fromValue(PaymentState.APPROVED.name()));
+  }
+}
