+++ payment/src/main/java/com/vroong/payment/adapter/in/rest/mapper/PaymentMapper.java
@@ -1,26 +1,26 @@
 package com.vroong.payment.adapter.in.rest.mapper;
 
 import com.vroong.payment.domain.Payment;
-import com.vroong.template.rest.MoneyDto;
-import com.vroong.template.rest.PaymentDto;
-import com.vroong.template.rest.PaymentStateDto;
-import org.jetbrains.annotations.NotNull;
+import com.vroong.payment.rest.MoneyDto;
+import com.vroong.payment.rest.PaymentDto;
+import com.vroong.payment.rest.PaymentStateDto;
 
 public class PaymentMapper {
-    public PaymentDto toDto(Payment entity) {
-        return new PaymentDto().paymentId(entity.getId())
-                .cardNumber(entity.getCardNumber())
-                .approvalNumber(entity.getApprovalNumber())
-                .amount(getMoneyDto(entity))
-                .state(getState(entity));
-    }
 
-    @NotNull
-    private PaymentStateDto getState(Payment entity) {
-        return PaymentStateDto.fromValue(entity.getState().name());
-    }
+  public PaymentDto toDto(Payment entity) {
+    return new PaymentDto()
+        .paymentId(entity.getId())
+        .cardNumber(entity.getCardNumber())
+        .approvalNumber(entity.getApprovalNumber())
+        .amount(getMoneyDto(entity))
+        .state(getState(entity));
+  }
 
-    private MoneyDto getMoneyDto(Payment entity) {
-        return new MoneyDto().value(entity.getAmount().getValue());
-    }
+  private PaymentStateDto getState(Payment entity) {
+    return PaymentStateDto.fromValue(entity.getState().name());
+  }
+
+  private MoneyDto getMoneyDto(Payment entity) {
+    return new MoneyDto().value(entity.getAmount().getValue());
+  }
 }
