+++ payment/src/main/java/com/vroong/payment/config/Constants.java
@@ -2,7 +2,7 @@ package com.vroong.payment.config;
 
 public class Constants {
 
-  public static final String PROJECT_NAME = "template";
+  public static final String PROJECT_NAME = "payment";
   public static final String SYSTEM_ACCOUNT = "system";
   public static final String UNKNOWN_USER_ID = "00000000-0000-0000-0000-000000000000";
   public static final String V1_MEDIA_TYPE = "application/vnd.vroong.private.v1+json";
@@ -43,8 +43,8 @@ public class Constants {
   }
 
   public static class LogKey {
-    public static final String HTTP_INBOUND_LOGGER = "com.vroong.template.http.api";
-    public static final String HTTP_OUTBOUND_LOGGER = "com.vroong.template.http.external";
+    public static final String HTTP_INBOUND_LOGGER = "com.vroong.payment.http.api";
+    public static final String HTTP_OUTBOUND_LOGGER = "com.vroong.payment.http.external";
     public static final String REQUEST = "request";
     public static final String RESPONSE = "response";
     public static final String USER_NAME = "userName";
