+++ payment/src/main/java/com/vroong/payment/domain/Payment.java
@@ -1,12 +1,18 @@
 package com.vroong.payment.domain;
 
 import com.vroong.shared.Money;
+import javax.persistence.Entity;
+import javax.persistence.EntityListeners;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.ToString;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.*;
 
 @Entity
 @Table(name = "payments")
@@ -15,13 +21,15 @@ import javax.persistence.*;
 @ToString
 @EqualsAndHashCode(of = "id")
 public class Payment {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-    private String cardNumber;
-    private String approvalNumber;
-    private Money amount;
 
-    @Enumerated(EnumType.STRING)
-    private PaymentState state;
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  private String cardNumber;
+  private String approvalNumber;
+  private Money amount;
+
+  @Enumerated(EnumType.STRING)
+  private PaymentState state;
 }
