+++ payment/src/main/java/com/vroong/payment/domain/PaymentState.java
@@ -1,8 +1,8 @@
 package com.vroong.payment.domain;
 
 public enum PaymentState {
-    WAITING,
-    APPROVED,
-    REJECTED,
-    CANCELED
+  WAITING,
+  APPROVED,
+  REJECTED,
+  CANCELED
 }
