+++ payment/src/main/resources/config/bootstrap.yml
@@ -1,3 +1,3 @@
 spring:
   application:
-    name: template
+    name: payment
