+++ payment/src/main/resources/db/schema.sql
@@ -25,10 +25,10 @@ CREATE TABLE IF NOT EXISTS persistent_events
 
 CREATE TABLE IF NOT EXISTS payments
 (
-    id bigint(20) NOT NULL AUTO_INCREMENT,
-    card_number varchar(40) NOT NULL,
+    id              bigint(20)  NOT NULL AUTO_INCREMENT,
+    card_number     varchar(40) NOT NULL,
     approval_number varchar(40) NULL DEFAULT NULL,
-    amount decimal DEFAULT 0,
-    state varchar(10) NOT NULL,
+    amount          decimal          DEFAULT 0,
+    state           varchar(10) NOT NULL,
     PRIMARY KEY (id)
 );
