+++ payment/src/main/resources/sentry.properties
@@ -1,2 +1,2 @@
-stacktrace.app.packages=com.vroong.template
+stacktrace.app.packages=com.vroong.payment
 release=#project.version#
