+++ payment/src/main/resources/swagger/api.yml
@@ -1,7 +1,7 @@
 openapi: '3.0.3'
 
 info:
-  title: 'template'
+  title: 'payment'
   version: 0.0.1
 
 servers:
