+++ payment/src/test/java/com/vroong/payment/PaymentApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class TemplateApplicationTests {
+class PaymentApplicationTests {
 
 	@Test
 	void contextLoads() {
