+++ payment/src/test/resources/config/application.yml
@@ -1,6 +1,6 @@
 spring:
   application:
-    name: template
+    name: payment
 
   profiles:
     active: test
